Tried to clean up Offer.php (desc)
Offer.php is impossible to clean up. It needs to be rewritten from a scratch. Wonderful idea someone had to throw everything possible in ONE function. I am done with trying to redo this "class". Every time I change a single line everything breaks. I am leaving it as it is.